# Reduce shard count from 5 -> 2