@chris-erickson 2019-05-29
@nicpottier let me know what you think.  Not sure how you build a release, but let me know if I need to do anything else.