@nicpottier 2019-05-30
Excellent, thanks for the detailed experiment!