@nicpottier 2019-05-30
FYI for the future, the ChangeLog changes are generated off of commits when we create new releases, so no need to update those.